+++ lib/db/src/schema/index.ts
@@ -21,3 +21,4 @@ export * from "./initiatives";
 export * from "./validations";
 export * from "./calculations";
 export * from "./providerTests";export * from "./backlog";
+export * from "./environment";